Commit 7 - Notification and Sound
Added Notification when user leaves the activity that contains the timer, notification is dismissed when user reenters that activity

Added timer to notification; timer from activity is displayed in notification

Added start/stop button in notification to start or stop the timer in the activity

Added ding sound when user completes a set and timer is reset

Added the ability to reset current set on one click and current exercise on long click of the reset button

Added ability to skip current set on one click and current exercise on long click of skip button

Scaled text of XML title TextViews so they weren't smaller than the values they represented (tidying up the UI)

+++ Workout/app/src/main/java/com/example/stopwatch/models/SetModel.java
@@ -2,11 +2,11 @@ package com.example.stopwatch.models;
 
 public class SetModel {
     public int reps;
-    public int weight;
+    public double weight;
     public int time;
 
     public SetModel(){}
-    public SetModel(int reps, int weight, int time) {
+    public SetModel(int reps, double weight, int time) {
         this.reps = reps;
         this.weight = weight;
         this.time = time;
